requirements.txt
feedparser==6.0.10
anthropic==0.18.1
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
